fix hotel rack

+++ app/Http/Controllers/Api/V1/Supplier/RoomController.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\ApiController;
 use App\Inside\Constants;
 use App\Inside\Helpers;
 use App\Room;
+use App\RoomEpisode;
 use Illuminate\Http\Request;
 use App\Http\Requests;
 use Illuminate\Support\Facades\DB;
@@ -326,6 +327,11 @@ class RoomController extends ApiController
             throw new ApiException(
                 ApiException::EXCEPTION_NOT_FOUND_404,
                 'plz check your id'
+            );
+        if (RoomEpisode::where('hotel_id', $id)->exists())
+            throw new ApiException(
+                ApiException::EXCEPTION_NOT_FOUND_404,
+                'کاربر گرامی هتل مورد نظر دارای سانس می باشد.'
             );
         Room::where('id', $id)->delete();
         return $this->respond(["status" => "success"]);
